requirements.txt
Flask
mysql-connector-python



 CREATE DATABASE farmer_registration1;
  USE farmer_registration1;
  CREATE TABLE problems1 (
    id INT AUTO_INCREMENT PRIMARY KEY,
    name VARCHAR(100) NOT NULL,
    contact VARCHAR(100) NOT NULL,
    crop VARCHAR(100),
    issue TEXT NOT NULL,
    additional_info TEXT
);
